refactor(playground): reorganize mdx examples into separate files
Move mdx examples from a single file to individual files for better maintainability
Update imports to reference new file structure
