+++ apps/playground/components/layout/ExampleSheet.tsx
@@ -4,7 +4,7 @@ import { useState } from 'react'
 import { Menu, Github, Lightbulb } from 'lucide-react'
 import { Button } from '../ui/button'
 import { Sheet, SheetContent, SheetHeader, SheetTitle, SheetTrigger } from '../ui/sheet'
-import { mdxExamples } from '@/data/examples'
+import { mdxExamples } from '@/data'
 import { useEditorStore } from '@/store/editorStore'
 import { ToggleMode } from '../shared/ToggleMode'
 import Link from 'next/link'
