+++ apps/playground/data/2-basic-markdown.ts
@@ -1,0 +1,38 @@
+export const basicMarkdown = {
+  id: 'basic-markdown',
+  title: 'Basic Markdown',
+  description: 'Standard markdown syntax with headings, lists, and formatting',
+  content: `# Welcome to MDX Playground
+
+This is a paragraph with **bold text** and *italic text*.
+
+## Features
+
+- Live preview with manual refresh
+- Support for markdown syntax
+- Component integration
+- Code highlighting
+
+### Formatting Examples
+
+You can use \`inline code\` and also create links like [MDX Documentation](https://mdxjs.com).
+
+> This is a blockquote with some inspiring text.
+> It can span multiple lines.
+
+---
+
+## Lists
+
+### Unordered List
+- First item
+- Second item
+  - Nested item
+  - Another nested item
+- Third item
+
+### Ordered List
+1. First step
+2. Second step
+3. Third step`,
+}
