+++ apps/playground/data/4-components.ts
@@ -1,0 +1,53 @@
+export const components = {
+  id: 'components',
+  title: 'Rich Content',
+  description: 'Images, links, and rich markdown content',
+  content: `# Rich Markdown Content
+
+This example shows various markdown features for rich content creation.
+
+## Images
+
+![Placeholder Image](https://via.placeholder.com/400x200/0ea5e9/ffffff?text=Sample+Image)
+
+*Caption: This is a sample image with caption*
+
+## Links and References
+
+Visit [MDX Documentation](https://mdxjs.com) for more information.
+
+You can also use reference-style links like [this one][1].
+
+[1]: https://mdxjs.com
+
+## Emphasis and Formatting
+
+- **Bold text** for emphasis
+- *Italic text* for subtle emphasis
+- ***Bold and italic*** for strong emphasis
+- ~~Strikethrough~~ for corrections
+- \`inline code\` for technical terms
+
+## Keyboard Shortcuts
+
+Press <kbd>Ctrl</kbd> + <kbd>C</kbd> to copy
+Press <kbd>Ctrl</kbd> + <kbd>V</kbd> to paste
+Press <kbd>Ctrl</kbd> + <kbd>S</kbd> to save
+
+## Horizontal Rules
+
+Content above the rule
+
+---
+
+Content below the rule
+
+## Nested Lists
+
+1. First level item
+   - Second level bullet
+   - Another second level
+     1. Third level numbered
+     2. Another third level
+2. Back to first level`,
+}
